Refactors ConfigueHacc
