+++ app/imports/ui/components/administrator/ChallengeAdminWidget.jsx
@@ -7,7 +7,6 @@ import { removeItMethod } from '../../../api/base/BaseCollection.methods';
 import { Challenges } from '../../../api/challenge/ChallengeCollection';
 import { COMPONENT_IDS } from '../../testIDs/componentIDs';
 
-/** Renders a single row in the table. See pages/ManageHaccWidget.jsx. */
 const ChallengeAdminWidget = ({ challenge }) => {
   const removeItem = (docID) => {
     swal({
