--- app/imports/ui/pages/administrator/ConfigureHaccPage.jsx
@@ -1,9 +1,0 @@
-import React from 'react';
-import withAllSubscriptions from '../../layouts/AllSubscriptionsHOC';
-import ManageHaccWidget from '../../components/administrator/ManageHaccWidget';
-
-const ConfigureHaccPage = () => (
-    <ManageHaccWidget/>
-);
-
-export default withAllSubscriptions(ConfigureHaccPage);
